refactor(ui): extract resource findings columns to separate file
