+++ ui/components/resources/table/index.ts
@@ -1,4 +1,5 @@
 export * from "../skeleton/skeleton-table-resources";
 export * from "./column-resources";
 export * from "./resource-detail";
+export * from "./resource-findings-columns";
 export * from "./resources-table-with-selection";
